Extracted Prometheus parser to parsers
