+++ plugins/inputs/prometheus/prometheus.go
@@ -15,6 +15,7 @@ import (
 	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/plugins/common/tls"
 	"github.com/influxdata/telegraf/plugins/inputs"
+	parser "github.com/influxdata/telegraf/plugins/parsers/prometheus"
 )
 
 const acceptHeader = `application/vnd.google.protobuf;proto=io.prometheus.client.MetricFamily;encoding=delimited;q=0.7,text/plain;version=0.0.4;q=0.3,*/*;q=0.1`
@@ -321,10 +322,14 @@ func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error 
 		return fmt.Errorf("error reading body: %s", err)
 	}
 
-	if p.MetricVersion == 2 {
-		metrics, err = ParseV2(body, resp.Header)
-	} else {
-		metrics, err = Parse(body, resp.Header)
+	isprotobuf, err := parser.IsProtobuf(resp.Header)
+
+	if err == nil {
+		parser := parser.Parser{
+			MetricVersion: p.MetricVersion,
+			Protobuf:      isprotobuf,
+		}
+		metrics, err = parser.Parse(body)
 	}
 
 	if err != nil {
